moves location form into its own component and then render component in profile when add button is clicked
